secboot: define RecoveryKey even when using nosecboot
